site/requirements.txt
django
djangorestframework
django-configurations
pylint
mysqlclient
django-cors-headers